factored out run_cuda_eval, so it can be used independently of modal
