+++ src/discord-cluster-manager/modal_runner.py
@@ -1,10 +1,9 @@
-import signal
 import subprocess
-from contextlib import contextmanager
 from typing import Optional
 
-from consts import CUDA_FLAGS, MODAL_CUDA_INCLUDE_DIRS, MODAL_PATH
+from consts import MODAL_PATH
 from modal import App, Image, Mount
+from timeout import TimeoutException, timeout
 
 # Create a stub for the Modal app
 # IMPORTANT: This has to stay in separate file or modal breaks
@@ -44,29 +43,6 @@ cuda_image = (
         # "cd /ThunderKittens && pwd && python setup.py install",
     )
 )
-
-
-class TimeoutException(Exception):
-    pass
-
-
-@contextmanager
-def timeout(seconds: int):
-    """Context manager that raises TimeoutException after specified seconds"""
-
-    def timeout_handler(signum, frame):
-        raise TimeoutException(f"Script execution timed out after {seconds} seconds")
-
-    # Set up the signal handler
-    original_handler = signal.signal(signal.SIGALRM, timeout_handler)
-    signal.alarm(seconds)
-
-    try:
-        yield
-    finally:
-        # Restore the original handler and disable the alarm
-        signal.alarm(0)
-        signal.signal(signal.SIGALRM, original_handler)
 
 
 def run_pytorch_script(  # noqa: C901
@@ -142,107 +118,6 @@ def run_pytorch_script(  # noqa: C901
         return f"Error executing script: {str(e)}", 0.0
     finally:
         tmp_files = ["eval.py", "reference.py", "train.py"]
-        for f in tmp_files:
-            if os.path.exists(f):
-                os.remove(f)
-
-
-def run_cuda_script(  # # noqa: C901
-    script_content: str,
-    reference_content: str = None,
-    submission_content: str = None,
-    timeout_seconds: int = 600,
-    arch: int = None,
-) -> tuple[str, float]:
-    """
-    Executes the provided CUDA kernel in an isolated environment with a timeout
-
-    Args:
-        script_content: The CUDA script containing the GPU kernel
-        reference_content: The (optional) reference code, used for leaderboards.
-        submission_content: The (optional) submission code, used for leaderboards.
-        timeout_seconds: Maximum execution time in seconds (default: 600 seconds)
-        arch: The arch code for the compute/sm versions.
-
-    Returns:
-        tuple[str, float]: (Kernel output, execution time in milliseconds)
-
-    NOTE: Modal execution time is not programmatically accessible, so we manually calculate it
-    """
-    import os
-    import subprocess
-    import time
-
-    try:
-        with timeout(timeout_seconds):
-            # Check CUDA is available and installed correctly
-            print("[CUDA Env Check]")
-            try:
-                # these check cuda compiler is also available
-                subprocess.run(["nvcc", "--version"], check=True)
-                subprocess.run(["which", "nvcc"], check=True)
-            except Exception:
-                return "nvcc not found.", 0.0
-
-            ARCH = f"-gencode=arch=compute_{arch},code=sm_{arch}"
-            NVCC_FILES = "eval.cu"
-            # Write submission files to directory
-            if reference_content is not None:
-                with open("reference.cuh", "w") as f:
-                    f.write(reference_content)
-
-            if submission_content is not None:
-                with open("train.cuh", "w") as f:
-                    f.write(submission_content)
-
-            with open("eval.cu", "w") as f:
-                f.write(script_content)
-
-            execution_start_time = time.perf_counter()
-            compile_process = subprocess.run(
-                ["nvcc"]
-                + CUDA_FLAGS
-                + MODAL_CUDA_INCLUDE_DIRS
-                + [ARCH, NVCC_FILES, "-o", "eval.out"],
-                capture_output=True,
-                text=True,
-            )
-
-            if compile_process.returncode != 0:
-                raise RuntimeError(
-                    "CUDA compilation failed with return code "
-                    + f"{compile_process.returncode}:\n{compile_process.stderr}"
-                )
-
-            run_process = subprocess.run(["./eval.out"], capture_output=True, text=True)
-            execution_end_time = time.perf_counter()
-
-            print("run process stdout", run_process.stdout)
-
-            score = None
-            for line in run_process.stdout.splitlines():
-                if line.startswith("score:"):
-                    score = float(line.split(":")[1].strip())
-                    break
-
-            if score is None:
-                execution_end_time = time.perf_counter()
-                score = execution_end_time - execution_start_time
-                return (
-                    "check_implementation failed"
-                    if "check_implementation failed" in run_process.stdout
-                    else None,
-                    score,
-                )  # To make sure error is thrown on LB
-
-            return run_process.stdout, score
-
-    except TimeoutException as e:
-        return f"Timeout Error: {str(e)}", 0.0
-    except Exception as e:
-        return f"Error executing script: {str(e)}", 0.0
-    finally:
-        tmp_files = ["reference.cuh", "train.cuh", "eval.cu", "eval.out"]
         for f in tmp_files:
             if os.path.exists(f):
                 os.remove(f)
