+++ src/discord-cluster-manager/timeout.py
@@ -1,0 +1,24 @@
+import signal
+from contextlib import contextmanager
+
+
+class TimeoutException(Exception):
+    pass
+
+
+@contextmanager
+def timeout(seconds: int):
+    """Context manager that raises TimeoutException after specified seconds"""
+
+    def timeout_handler(signum, frame):
+        raise TimeoutException(f"Script execution timed out after {seconds} seconds")
+
+    # Set up the signal handler
+    original_handler = signal.signal(signal.SIGALRM, timeout_handler)
+    signal.alarm(seconds)
+
+    try:
+        yield
+    finally:
+        # Restore the original handler and disable the alarm
+        signal.alarm(0)
